Update XIPacket Definitions
Makes a slight adjustment to the packet header definiton.

Also completes the implementation of the XIPacket and Header.

The XIPacket/Header can now be represented by a string.

A XIPacket/Header can now be constructed by a given string buffer.

Adds sequence number to packet header definition.
  Update is as follows:
    XIPacketHeader:
      Second Byte:
  	type          : 2b
	src           : 1b
	dst           : 1b
	seq           : 2b
	payload (MSB) : 2b
	-------------------------
	                8b

Adds test cases to check all possible non default args (with the
exception of payload) for XIPacket and XIPacketHeader.